# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Doh Eain - Portal",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/items": {
            "get": {
                "tags": [
                    "Item"
                ],
                "summary": "List",
                "description": "Get Item List",
                "operationId": "itemList",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "per_page",
                                    "page",
                                    "sort_by"
                                ],
                                "properties": {
                                    "q": {
                                        "type": "string"
                                    },
                                    "per_page": {
                                        "type": "integer"
                                    },
                                    "page": {
                                        "type": "integer"
                                    },
                                    "sort_by": {
                                        "type": "string"
                                    },
                                    "sort_desc": {
                                        "type": "boolean"
                                    },
                                    "type": {
                                        "type": "string"
                                    },
                                    "category_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "post": {
                "tags": [
                    "Item"
                ],
                "summary": "Create",
                "description": "Create Item",
                "operationId": "itemCreate",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "category_id",
                                    "unit_id",
                                    "charge",
                                    "type"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "category_id": {
                                        "type": "integer"
                                    },
                                    "unit_id": {
                                        "type": "integer"
                                    },
                                    "charge": {
                                        "type": "number"
                                    },
                                    "type": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/items/{id}": {
            "get": {
                "tags": [
                    "Item"
                ],
                "summary": "Get",
                "description": "Get Item",
                "operationId": "itemGet",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "put": {
                "tags": [
                    "Item"
                ],
                "summary": "Update",
                "description": "Update Item",
                "operationId": "itemUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "category_id": {
                                        "type": "integer"
                                    },
                                    "unit_id": {
                                        "type": "integer"
                                    },
                                    "charge": {
                                        "type": "number"
                                    },
                                    "type": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Item"
                ],
                "summary": "Delete",
                "description": "Delete Item",
                "operationId": "itemDelete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/options/{object}": {
            "get": {
                "tags": [
                    "Common"
                ],
                "summary": "Get",
                "description": "Get id and name data array in database table. Parameter {object} is the table name [user, role, unit, category, item, patient]",
                "operationId": "optionGet",
                "parameters": [
                    {
                        "name": "object",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}